Time: 3 ms (45.34%), Space: 8.5 MB (67.54%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         if (x<0) return false;
-        double rev = 0;
+        long long rev = 0;
         int n = x;
         while (n != 0) {
             int digit = n % 10;
